+++ tests/unit/TestExecutionLimit.pas
@@ -13,6 +13,9 @@ type
   public
     [Test]
     procedure ExecutionLimit_StopsAfterN;
+
+    [Test]
+    procedure ExecutionCount_SkipIfRunning_OnlyCountsExecutions;
   end;
 
 implementation
@@ -49,6 +52,48 @@ begin
     Assert.AreEqual(2, TInterlocked.CompareExchange(Count, 0, 0));
   finally
     Cron.Free;
+  end;
+end;
+
+procedure TTestExecutionLimit.ExecutionCount_SkipIfRunning_OnlyCountsExecutions;
+var
+  lCron: TmaxCron;
+  lEvt: TmaxCronEvent;
+  lStarted: TEvent;
+  lGate: TEvent;
+  lWaitRes: TWaitResult;
+begin
+  lCron := TmaxCron.Create(ctPortable);
+  try
+    lStarted := TEvent.Create(nil, True, False, '');
+    lGate := TEvent.Create(nil, True, False, '');
+    try
+      lEvt := lCron.Add('SkipCount');
+      lEvt.EventPlan := '* * * * * * * 0';
+      lEvt.InvokeMode := imThread;
+      lEvt.OverlapMode := omSkipIfRunning;
+      lEvt.OnScheduleProc :=
+        procedure(Sender: TmaxCronEvent)
+        begin
+          lStarted.SetEvent;
+          lGate.WaitFor(3000);
+        end;
+      lEvt.Run;
+
+      lCron.TickAt(lEvt.NextSchedule);
+      lWaitRes := lStarted.WaitFor(2000);
+      Assert.AreEqual(TWaitResult.wrSignaled, lWaitRes);
+
+      lCron.TickAt(lEvt.NextSchedule);
+      Assert.AreEqual(1, Integer(lEvt.NumOfExecutionsPerformed));
+
+      lGate.SetEvent;
+    finally
+      lGate.Free;
+      lStarted.Free;
+    end;
+  finally
+    lCron.Free;
   end;
 end;